fix: extract mutation into lib
