+++ apps/journeys-admin/src/libs/useUpdateJourneyHostMutation/index.ts
@@ -1,0 +1,1 @@
+export { useUpdateJourneyHostMutation } from './useUpdateJourneyHostMutation'
